Test the tier contract once, in the core that owns it
SonarCloud failed the gate on duplication, and it was right: the startup
validation test sat verbatim in both package projects, twenty identical
lines.

Deduplicating by extracting a helper would have missed the point. What
those tests check belongs to AddCustodian and HoldKeysInCustodian, which
live in the core and behave the same whichever custodian is registered,
so asserting them through Vault and again through Azure was testing the
core through two arbitrary doors. They now run once against a stub
custodian, alongside the ordering test that was already there.

Each package keeps only what is its own: that its typed client is the
custodian, that the tier call installs the provider, and for Vault, the
Transit base address and auth header.
