+++ Abblix.Oidc.Server.Vault.UnitTests/AddVaultCustodianTests.cs
@@ -30,9 +30,9 @@ using Xunit;
 namespace Abblix.Oidc.Server.Vault.UnitTests;
 
 /// <summary>
-/// Verifies the wiring performed by <c>AddVaultCustodian</c>: the typed Transit client is pointed at the mount
-/// with its auth header and registered as the custodian, the tier call chained onto it installs the key provider,
-/// and omitting that tier call fails at startup instead of silently serving the static keys from the options.
+/// Verifies what is specific to <c>AddVaultCustodian</c>: the typed Transit client is pointed at the mount with
+/// its auth header and registered as the custodian, and the tier call chained onto it installs the key provider.
+/// What the tier call itself enforces is custodian-agnostic and lives in the core's own wiring tests.
 /// </summary>
 public class AddVaultCustodianTests
 {
@@ -78,47 +78,6 @@ public class AddVaultCustodianTests
         // The Transit client itself serves as the external key custodian, and the provider publishes its keys.
         Assert.IsType<VaultTransitClient>(provider.GetRequiredService<IKeyCustodian>());
         Assert.IsType<ExternalKeysProvider>(provider.GetRequiredService<IAuthServiceKeysProvider>());
-    }
-
-    [Fact]
-    public void StartupValidationFails_WhenTheTierIsNeverChosen()
-    {
-        var services = new ServiceCollection();
-        AddCustodian(services);
-        services.AddSingleton(TimeProvider.System);
-
-        using var provider = services.BuildServiceProvider();
-
-        // The host runs the startup validators before it starts the hosted service that opens the HTTP port, so
-        // this is the failure a misconfigured deployment actually meets: no port, no token, no silent fallback to
-        // the static keys of OidcOptions.
-        var validator = provider.GetRequiredService<IStartupValidator>();
-
-        var error = Assert.Throws<OptionsValidationException>(validator.Validate);
-        Assert.Contains("HoldKeysIn", Assert.Single(error.Failures));
-    }
-
-    [Fact]
-    public void StartupValidationPasses_WhenTheTierIsChosen()
-    {
-        using var provider = Configure().BuildServiceProvider();
-
-        provider.GetRequiredService<IStartupValidator>().Validate();
-    }
-
-    [Fact]
-    public void KeyProviderStillGuards_WhenTheTierIsNeverChosen()
-    {
-        var services = new ServiceCollection();
-        AddCustodian(services);
-        services.AddSingleton(TimeProvider.System);
-
-        using var provider = services.BuildServiceProvider();
-        var keysProvider = provider.GetRequiredService<IAuthServiceKeysProvider>();
-
-        // The second line, for a host that resolves keys without a host lifetime to run the startup validation.
-        var error = Assert.Throws<InvalidOperationException>(() => keysProvider.GetSigningKeys());
-        Assert.Contains("HoldKeysIn", error.Message);
     }
 
     [Fact]
